修改

+++ src/main/java/com/aliabao/springbootspringdatajpa/respository/UserRespository.java
@@ -9,6 +9,6 @@ import java.io.Serializable;
 //JpaRepository提供的简单数据操作接口
 //JpaSpecificationExecutor(SpringDataJPA提供的复杂查询接口)
 //SpringDataJPA内部使用了类代理的方式让继承了他的接口都以spring管理的Bean的形式存在
-public interface UserRespository extends JpaRepository<User,Long>,JpaSpecificationExecutor<User>,Serializable {
+public interface UserRespository extends JpaRepository<User,Integer>,JpaSpecificationExecutor<User>,Serializable {
 
 }
